[nova] Handle changed shellinabox URLs
Ironic changed the URLs returned for the shellinabox console from
containing no path to including information about the node, timeout and
a digest in the path. Since we have to include the token in the URL
during a redirect and we took any string as a token, this broke the
shellinabox console in Nova.

To fix this, we now parse explicitly UUIDs from the URL instead of any
string. Additionally, we also have to react to URLs with a path instead
of only on "/".