[nova] Automatically confirm resize after three days
Users often forget to confirm a resize after the vm has been
successfully resized.

Automatically confirming brings the VMs back to a standard state and
clean up the copy of the root disk.

The time-frame was chosen to be a weekend plus one day to be hopefully
as short as possible without messing up any manual process going on,
which might get delayed over a weekend.