[nova] Support and configure weights for PreferSameShardOnResizeWeigher
We want VMs to not move between shards on resize if possible, because
this takes a long time for VMs having volumes attached because of the
amount of data needing migration. We already prefer the same host on
resize, but if that's not possible e.g. because of resizing to HANA or
away from a small VM, the same shard should be used.

To accommodate that, we set the weight for the
`PreferSameShardOnResizeWeigher` to 200. This still prefers the same
host if possible, which is advantageous because of less ephemeral disk
copying/duplication, but prefer the same shard over the right RAM class.
E.g., the right RAM class with more space in another shard would now,
with weight 100 on the `HvRamClassWeigher` have a weight of 1xx, while
the weight of the wrong RAM class in the same shard would now be 2xx and
by that overpower the RAM class in another shard.