[neutron] Add metric that exposes the last heartbeat's timestamp
Exposing the seconds since the last timestamp as seen from the database
adds some complexity as prometheus will carry that value as it was on
time the exporter run. This makes it necessary to add the time passed
when the exporter was last run to get the actual value.

Just exposing a unix timestamp makes it much easier.