[manila] rate-limiting "update/access_list" as read action
The API for getting access rules is POST prior to 2.44. So clients using
lower API version hit the write limits and get 429 "Too Manay Resquests"
errors. We modify the group rule for rate limit middleware, so that such
requests are classified as "read" action. Note the read group rules are
moved to the front of write group rules and have higher priorities,
because the rules are parsed from top to bottom in rate limit
middleware.