keystone: controls for delayed assignment cleanup
Immediately deleting role assignments is sometimes not desirable. People might
forget to renew their ldap group membership, be on vacation or have other
temporary short-term leave. Another case is a temporary unavalability of
CAM resulting in users not being visible in keystone.

If the role assignment is gone, projects are left unattended, and it is not
desirable.