[PX] Add chart to temporary check on all nodes if PX vlans are present
We now have had multiple occasions when a pod was rescheduled to a host
that had either a host misconfiguration or the connected switch was
misconfigured. This resulted in the VLAN not being available and hence
the route servers being unreachable. We want to use this deployment as a
site survey to find out were this vlan is misconfigured. As IPs are a
limited resource in the peering nets, we rely on ARP/l2 only.