[neutron] Alerts for rate limit hits
Introducing two alerts based on the number of global (project independent) and local
(project-based) rate limit hits. Whenever a limit exceeds 10% of the overall
requests for 5 minutes the alert triggers. The project independent based alert compares
the tuple of action (create, read, etc.) and URI hitting the limit in relation to the overall number of requests of that type.
Comparing the local vs global based alerts reveals that the first additionally groups by the project_id.

Both alerts are based on metrics provided by the
openstack-watcher-middleware and openstack-rate-limit-middleware. The
first provides the information how many requests are being send. The
latter provides the information of limit hits.