[Nova] Move OpenstackNovaComputeIdle into Thanos
We want the alert to only fire if the service is not disabled. Our main
metric comes from prometheus-kubernetes, while the
`nova_compute_service_status` metric containing the service status (0
for disabled, 1 for enabled) comes from prometheus-openstack. Both
metrics are available in `metrics.metal.$region.cloud.sap`, so we move
the rule into that Thanos instance.

To support Thanos, we need to extend our alert template. We detect
thanos by the prefix "metrics-" for the directory. The docs [0] state,
that we need to set a different label and then everything should work.

[0] https://operations.global.cloud.sap/docs/architecture/monitoring/alerts/#thanos-rulesthanosrules