nova/metrics: Fix empty instances gauge
The query used the disused and now mostly-empty `instance_types`
table, which made the metric report no values (or only a very few
ancient instances in some regions).

Use instance_extra instead and parse the flavor name out of the JSON.
The query is a bit more expensive this way (it doubled in time from
the query without any join).