[cc-rbac] Introducting namespacedSupporterGroups
Add support for namespace sensitive supporter roles. For PX we need to
give team members with the 'CC Converged Cloud Partner PX' CAM role
limited access to the cluster in the PX name space. They should be able
to read pods, deployments, configmaps, and kubectl exec. For that the
supporter cluster role is suitable.

To be able to limit it to the PX namespace we add the
namespacedSupporterGroups to enable fine grained access.