[RabbitMQ] add docker file to build rabbitmq image
Upstream RabbitMQ image depending on ubuntu 22.04 upstream images, it
does not perform an Os upgrade while building the image which may cause
situations where the base image is few days old and does not contain the
latest updates.
for that adding a Dockerfile to perform an OS upgrade on the RabbitMQ
upstream image.

Change-Id: Ie8dcc7844186bbc388e9bc146d6411070a4302e9