Improve empty port-selector and multiple subgroup bindings handling
In the driver we do now support no port-selectors, if a host never has
to go into baremetal mode. This commit will omit the port_selectors key
if they are unset.

The policy group was inferred to be the same name as the VPC profile's
first binding's name. This failed when our subgroups got multiple
bindings which lead to an empty string being rendered into the config.
This shouldn't go unnoticed so we prompt the user to supply an explicit
pc-group only if port selectors are set. Policy groups are not needed if
port-selectors are unset, as no disassembling of the VPC profile will
happen.