Require an explicit policy group if >1 bindings per subgroup
The policy group was inferred to be the same name as the VPC profile's
first binding's name. This failed when our subgroups got multiple
bindings which lead to an empty string being rendered into the config.
This shouldn't go unnoticed so we prompt the user to supply an explicit
pc-group.