[utils] Remove set_openstack_request_id snippet
We used this snippet with the `configuration-snippet` annotation in our
Ingress objects, but since `configuration-snippet` got disabled for
security-reasons, we do not need this snippet anymore. Instead, the
functionality is integrated into the ingress itself.

We bump the minor version here, so no projects are getting automatic
updates of the utils chart with the removed snippet.