[utils] Add linkerd helpers
While they might also seem to fit directly into the
common/linkerd-support chart, they don't, because the snippets would not
be helpful for all the charts consuming the linkerd-support chart.
Specifically those in common/ would not be able to access them if their
parent chart includes linkerd-support. Having the chart as extra
dependency is also not wanted. Therefore, we add these helper snippets
here to use them in the openstack charts.

There are 2 annotation helpers defined for which more documentation can
be found in the README of the linkerd-support chart.

Instead of having to put the 3-4 lines in every template, these snippets
allow us to add only one and also update it centrally.