[Nova] Add basic linkerd support
This adds the `linkerd-support` chart as dependency and also adds the
annotations to all `Deployment`, `Service` and `Ingress` objects. `Job`
objects are out of the picture for now, because they need more than just
annotations as we need to shut down linkerd for the pod to shut down.
There will be a follow-up to support these.

linkerd is disabled by default and needs to be enabled per region.

We depend on the `utils` chart for the annotation snippets and the
conditions so our templates stay more compact and thus better readable.