[Cinder] Update chart to include linkerd
This adds the `linkerd-support` chart as dependency and also adds the
annotations to all `Deployment`, `Service` and `Ingress` objects.

linkerd is disabled by default and needs to be enabled per region.

We depend on the `utils` chart for the annotation snippets and the
conditions so our templates stay more compact and thus better readable.