[openstack/utils] Common snippets for a trust-bundle
The trust-manager[^1] offers the capability to centrally manage
root certificates. This allows us to inject a root of trust with
kubernetes means during runtime instead of patching the images to
contain them during build-time.

We can revoke root-certificates in the same fashion.

The small side-effect is not having to patch images to contain our
company CA.

The bigger degree of flexibility also allows us to easily add a custom
CA for doing mutual tls authentication on the service level, adding
one option for removing credentials from the charts.

Defaulting to false for backwards compatibility.

[^1]: https://cert-manager.io/docs/trust/trust-manager/