[openstack/barbican] Use trust-bundle snippets, rename migration-job
The snippets replace the system ca trust-bundle with a centrally managed
one by trust-manager in k8s. This eliminates the need for patching the
images to contain the company internal CA.

The migration job is a helm hook, so the name doesn't need to be unique,
it will get re-created by helm.