octavia: fix jobs being left over after redeployment
The explanation in the deleted code comment that jobs cannot be
recreated was correct at the time, in commit 24c84622a6, because the job
was created as a regular manifest object instead of a hook object.

As of now, the job has a `hook-delete-policy: before-hook-creation`, so
job spec immutability is not a concern.

We found this issue while I was explaining the GkVulnerableImages check
to Benjamin Ludwig, and we saw that dozens of old images are kept from
being GC'd because these jobs are still referencing them.