Enable harakiri-graceful-signal
Send the SIGWINCH signal if a process gets shutdown in order to trigger
a guru meditation report creation. Requires harakiri_graceful_timeout to be > 0.

This behaviour is disabled per default.

The first harakiri attempt on a worker will be graceful if harakiri_graceful_timeout > 0,
then the worker has harakiri_graceful_timeout seconds to shutdown cleanly, otherwise
a second harakiri will trigger a SIGKILL