[designate-nanny] moving user seeds to designate
To enable seeding in all namespaces necessary, user seeding has moved
to the designate helm chart, to continue having a single source of
truth the nanny charts now read the designate values as well.

Enabling the nanny and the users used is now configured in value files
of designate centrally.