[manila] set default for reserved_snapshot_percentage
allows to create shares created from a snapshot subject to the
configured reserve.
This reserve can be lower than the general `reserved_percentage`,
thus allowing to create shares from snapshot (which usually and
at the time of writing always! end up on the same host) even when
creating new shares in general is not allowed at a certain high
utilization.