[neutron] Don't share netns between dhcp/bridge agent
The linux-bridge-agent and dhcp-agent used to share their /var/run/netns
directory with each other. This results in two things:

1. The network namespaces are generally not usable in the
   linux-bridge-agent
2. When the dhcp-agent container restarts, the netns files are still
   present, even when the namespaces themselves are gone

Generally, only the dhcp-agent interacts with the network namespaces.
The netns are managed by it and it also puts interfaces inside the
netns. The bridge-agent only puts interfaces provided to it into a
bridge. This is possible with and without the knowledge of the netns
from the dhcp-agent.

When the dhcp-agent restarts it can no longer recreate the network
namespaces, as the namespaces themselves are gone, but the files are
still there. It'll try to use these invalid files, but it won't work
anymore.

Therefore, to allow the dhcp-agent container to properly restart without
restarting the whole pod, we now remove this crossmount of the netns
directory.

Co-authored-by: Johannes Kulik <johannes.kulik@sap.com>