[Nova] Resolve OpenstackNovaComputeIdle alert
This alert was very generic as it checked the CPU consumption of a pod
against a hard limit. By now, we have services time and time again that
do not see enough load i.e. are idling most of the time. Therefore, this
alerts creates seemingly only false notifications by now.

Initially, we introduced this alert to find services that do not react
to RPC messages anymore. This hasn't occurred for a long time. If we
need this kind of monitoring again, we should look into the RPC ping
endpoint available from `oslo.messaging` and maybe into monitoring the
number of greenthreads or Lock waiters.

Even though there's no alert left for the "metrics-metal" source, we
keep it in place to make it easy to add new alerts again.