[Nova] nova-compute mounts cell-secrets from Secret
Instead of writing the database connection and rabbitmq transport_url
into the ConfigMap via the `vcenter-operator`, we instead create
configuration files in the `nova-etc` Secret for each cell. Each
Deployment of `nova-compute` created by the `vcenter-operator` then
mounts the appropriate Secret of the cell the `nova-compute` belongs to
into `/etc/nova/nova.conf.d/${CELL}-secrets.conf`.

This means, we only have to supply the cell names to the
`vcenter-operator` in the future, it doesn't have to read Secret
objects and it doesn't handle the actual secrets or references to them.