[Nova] Provide cell names via ConfigMap
The Nova helm-chart will now add an additional `nova-cells` ConfigMap
which contains a comma-separated list of cell names under the "cells"
key. This ConfigMap is provided for the `vcenter-operator` to provide
them to templates so templates can decide which secret should be
available for mounting (e.g. from `nova-etc` Secret).