[unbound][bind-rpz-proxy] -- switch the secrets from stringData: to data:
Apparently `stringData:` is broken by design [1]. We should avoid it.

The issue is that removing a stringData key will not remove it from the
k8s secret object. It will just get reset to an empty string.

Might be a security issue to some extend.

But, also, using such a secret as a volume will result in orphaned empty files.
That *might* break easily offended software.

Also, bumped the chart's `version` and `appVersion`.

[1] https://github.com/kubernetes/kubernetes/issues/89938#issuecomment-1033085353