[mariadb] move initdb script into secrets
Switch to use secret instead of configmap, to prevent having secrets in
plain text while templating the helmchart
configurator script added to configure mariadb usersout of a secret with
a job.
Because it is difficult to pass ENV variable to kubectl commands,the
script copies the templated SQL statements out of the secret to mariadb
and executes them from there.
kbuectl cp command does not work with files mounted out of secret, for
that the script copies the file to a temporary file on oder to copy it.
this version replaces the following values:
- initdb_configmap: with initdb_secret:
- custom_initdb_configmap: with custom_initdb_secret:

custom initdb SQL scripts must be provided as secret and assigned to
value custom_initdb_secret:

Change-Id: I3e5b9237bd51a002b0d5055818d7d14c85840393