[openstack/utils] Add "utils.linkerd.signal_stop_script"
This snippet contains the necessary commands to stop a `linkerd` sidecar
gracefully. It's included in the "utils.script.job_finished_hook" so
there should be no need to include it explicitly.

Since not all images have `curl` available - it's currently inside
loci-built images, but more as an oversight than by choice - we also
support sending a POST request to `linkerd` via `bash`. The POST is the
same action `linkerd-await` does to shut down `linkerd` [0]. We just
didn't want to add `linkerd-await` as dependency into our images and
commands.

[0] https://github.com/linkerd/linkerd-await/blob/880850995004e1c5b1f29fccdfdcbd4f45cb7ec2/src/main.rs#L245-L263