[Nova] Use Secret for username/password of keystone users
We add a "file" `keystoneauth-secrets.conf` to the `nova-etc` Secret and
move the username and password from the keystoneauth sections of
`nova.conf` into that file. From the `nova-etc` Secret, we mount it to
`/etc/nova/nova.conf.d/keystoneauth-secrets.conf` where it gets picked
up automatically by nearly all processes.

`openstack-agent-liveness` needs to be told about config files
explicitly, because it doesn't hand the project into `oslo.config` and
thus `oslo.config` can't automatically pick up the right configuration
directories. This could be changed as a follow-up to
`python-agentliveness`.

We switch the conductor for cell to a projected volume in the process to
align it with the other Deployments.