[Placement] Support linkerd in Jobs
We bump the `utils` chart and switch to "utils.script.job_finished_hook"
to include the commands to stop `proxysql` and `linkerd` at the end of
our Jobs.

With updating `utils`, we also get the trust-bundle only overwriting
"/etc/ssl/certs/ca-certificates.crt" instead of all of "/etc/ssl/certs".