[Nova] db-migrate reads secrets from file
Instead of templating the `transport_url`s and DB `connection`s of the
cells into the `nova-bin` ConfigMap, we read them from the config files
dynamically. To do this, we also have to introduce a `cell0.conf` into
`nova-etc` Secret.

We use Python's `ConfigParser()` to parse the cell configs because that
felt safer than using `grep` - because we previously had multiple
`connection` in a file that parsing the right one would get messy. By
now it feels like the Job would better be rewritten in Python directly.