[Nova] Move audit-middleware config to Secret
The `audit_middleware_notifications` section in `nova.conf` contained a
`transport_url` and thus secrets. Since we have to move all secrets to
Secrets, we now add a `audit-middleware.conf` via `nova.conf.d` into all
relevant pods.

In the process, we clean up all pods that do not need any audit-relevant
or even api-relevant configs. This makes it easier to spot places in
need for such changes in the future.