content-replication: remove restartPolicy from swift-http-import container
Restarting this container on failure does not work when linkerd
sidecars are injected, because linkerd-await in the container will
terminate the linkerd-proxy even when the s-h-i container fails. When
the s-h-i containers restarts, it will then not be able to talk to the
outside because its linkerd-proxy is gone.

However, since we have these pods wrapped in Jobs now, we can use the
native restart logic of Jobs instead to similar effect.