[nova] nova-compute-ironic to use the `nova` service user
nova-compute-ironic uses the ironic service user when talking to
the Ironic API. Since this user is also used by Ironic itself,
it's harder to coordinate password changes
(both Nova and Ironic need the new password at the same time).
Additionally, it's not clear that Nova is talking to Ironic
if the user is ironic.

With this change, the nova service user needs all the roles that
ironic user has, thus we're adding to the list:
- cloud_image_admin
- cloud_objectstore_admin