[utils] Render rabbitmq URLs with own helper
We do not want to depend on the `rabbitmq` chart to render our INI
sections. Therefore, we introduce `ini_section._transport_url` and
`utils.rabbitmq_url`. The first contains also the "transport_url = "
part necessary for the INI.

With `utils.rabbitmq_url` accepting a dict of parameters to render the
actualy URL, we can also use it to generate Nova's cell2 and the
auditmiddleware's URL instead of having to do that "manually". In
addition, there's only a single place that adds the
`resolve_secret_urlquery` part now.