[unbound] adding default values and guards in templates
when using the template-app to locally render the charts, it seems that
some values are missing. This leads to toYaml returning the literal string
'null' and breaking the yaml file as well as errors with the keyname
fields.

This commit adds some default values for these fields and an extra if
around the toYaml sections.