[ironic] fix creating secrets as pre-upgrade hook
We have the same issue as octavia in #6377 and #6382 and copy their
homework.
The migration job runs as pre-upgrade hook and it needs the
ironic-secret secret. Therefore the secret needs to be available during
the pre-upgrade hook time. It needs to have a lower weight than the
migration job, so that is executed before the migration job.
This needs to happen as an annotation