fix(ocb-qe): named template with unique names
see helm docs:
An important detail to keep in mind when naming templates: template names are global.
If you declare two templates with the same name, whichever one is loaded last will be the one used.
Because templates in subcharts are compiled together with top-level templates,
you should be careful to name your templates with chart-specific names.