[Ironic] Put the ipxe_config.template in a raw string
The `ipxe_config.template` is a jinja2 template interpreted by Ironic
code. Since both Helm and jinja2 use `{{` and `}}` as deliminator for
variables in the template, helm tries to interpret the jinja2 template.
To prohibit that, we "escaped" the `{{` and `}}` by telling helm to put
strings of `{{` and `}}` into the template.

With us not using much helm templating in this file, there's a better
readable way to do it: raw strings. We put basically the whole file into
a raw string, only ending the raw string where we need actual templating
from helm, starting it again right afterwards.

See also https://go.dev/ref/spec#String_literals and
https://pkg.go.dev/text/template#hdr-Examples