openstack/limes-auto: init at 0.1.0
The intention here is to deliver quota-overrides.yaml from the secrets
repo into prod without human intervention (only gated on validation).
Quota overrides are supposed to be maintained by the respective service
owners (e.g. the network-api team for Neutron and Octavia quotas), so
involving us in the deployment is an unnecessary complication.

This is a different deployment pattern than Limes itself, so it needs to
be located in a different chart.

This first step does not do anything on its own, but after this chart is
deployed, I will switch Limes to using the quota-overrides.json from
here instead of from its own chart.

To avoid human errors in automatic deployments, I will have additional
validation in CI. The limes-validation@Default user is being added for
this purpose.