[neutron] Increase network agent's memory limit
During a NFS filer downtime, with the neutron db scaled down, we noticed
that all our dhcp agents got OOM killed. The underling condition is
probably a memory leak that surfaces when neutron-rpc is not working
properly. We still haven't found it yet, but the quick
we-make-this-a-little-less-likely is to increase the memory limits. We
will still get OOM killed just a little later. The hope is that this
little extra time carries us through the filer downtime.