[manila] move OpenstackManilaSharePodScaledDown to thanos
inspired by a608b48fcb7990d955c838ab6c05e4a183fc05a9

We want to fire the alert only if the pod was a valid scheduling
target. The main metric comes from prometheus-kubernetes, while the
to be added metric (with the reserved percentage) comes from
prometheus-openstack.
Both metrics are available in `metrics.metal.$region.cloud.sap`,
so we move the rule into that Thanos instance.

To support Thanos, we need to extend our alert template. We detect
thanos by the prefix "metrics-" for the directory. The docs [0] state,
that we need to set a different label and then everything should work.

[0] https://operations.global.cloud.sap/docs/architecture/monitoring/alerts/#thanos-rulesthanosrules