[manila] reduce OpenstackManilaSharePodScaledDown in decom
For backends in decom we set reserved percentage to 100 to stop
scheduling.

So we fire the alert now only if this setting had been lower 7
days ago. We need to look in the past, because currently it won't be
reported - remember: the pod is scaled down and not reporting the
capabilities.

Eventually this means that we still get the alert for in_decom filers
that have been scaled down, just not after 12 hours, but after 7 days
and 12 hours instead.