[neutron] Skip ports for linker on network-agent
We skip outbound linkerd proxying for port 22,53,80,443 to better debug
networks from the network viewpoint of our network agents. Especially
port 22 will allow us to use the network agent as jumphost again.

The ports are added to the skip-inbound-ports as well due to the design
of linkerd and the network agents. Packets originating from the netns
will also traverse the linuxbridge-agent, where the packets will be
processed as "inbound" and will be forwarded to linkerd if we do not set
this option.