[ironic] Configure ESP with bootloader
This only affects deployments with a virtual-media interface.

It is a missing input for creating a bootable iso with the
ironic-python-agent inside.
The kernel and ramdisk are already specified, this just is the
bootloader used in the ISO created on the fly containing also
the token or certificate for authentication
This value can be overidden by setting a deploy-config value for
"bootloader".