[openstack-exporter] manila: improve ManilaAvailabilityZoneUsageHigh for small regions
In smaller regions we may not have 3 manila backend hosts per AZ.
But we want to make sure that at least every host has some capacity
left.

For the bigger regions we are okay with only having 3 hosts with
capacity, not all of them.

Created recording rules for better readability.