[Nova] Set sapphire_rapids_weight_multiplier
We want old general purpose flavors to land on old hardware if possible.
Therefore, we increase the weight of the SapphireRapidsWeigher - which
decreases their priority.

I decided to use the same value as we use for
`hv_ram_class_weight_multiplier` as they seem to be pretty similar in
what they want to achieve to me. I still want CUSTOM_DECOMMISSIONING to
be higher. Going lower than 100, we might end up on small sapphire
rapids (which can exist in theory) much easier.