nova/flavors: Minimal socket count for flavors >= 64vCPUs
We now want larger general purpose flavors to keep their cores
together instead of spreading them across sockets.

The underlying hypervisor topology is unrelated to the values that are
settable: Valid cores-per-socket numbers are the ones that result in
integer socket counts with the total number of CPUs, with a maximum of
64 cores.

E.g. with 96 vCPUs possible cores-per-socket values are:
48, 32, 24, 16, 12, 8, 6, 4, 3, 2, 1

The resulting socket-size can be larger than the socket-size of the
underlying hardware. The number of _NUMA nodes_ is influenced by the
underlying hardware, and the number of NUMA nodes can curiously enough
be greater than the number of sockets.