[Neutron] Disable eventlet patching in neutron-server
We set the environment variable DISABLE_EVENTLET_PATCHING, which is
honored by our own drivers and disables calls to
`eventlet.monkey_patch()`.

Additionally, we switch away from letting uWSGI use the WSGI file
`neutron-api`. The reason is, that `neutron.cmd.eventlet` uses
`eventlet.monkey_patch()` unconditionally - and is imported in the
`neutron-api` WSGI file. We switch instead to an `eval` setting for
uWSGI, because this makes it immediately visible what we're calling. The
code we're calling is also just three lines.

All of this only happens if we run with uWSGI. If we use the old
eventlet-wsgi server, we need the monkey-patching.