[Nova] Switch from raven-python to sentry-python
The old raven package is deprecated and out of maintenance since a
while. While it will stay mostly compatible, we can get a little more
information into Sentry, if we also configure the `BreadcrumbHandler`.
We want it to pick up any log line, so we configure it with DEBUG level.
This will only add breadcrumbs to an event that's actually sent in the
end. Which log lines get sent as event is configured by the level of the
`EventHandler`.