[mariadb-galera]add new value to indicate db name
Add a new value to indicate the name of the database needed to be used
in the connection string of the service.
Because the helm chart supports multiple database we added this value to
simplifies the logic on wraper helm-chart side to get the correct
database,for building the connection string to it.
That instead adding a flag to each database in mariadb.databases section
which indicates if the database connection relevant or not.
This value as well insures that only one value will be set and used.

Change-Id: I760714b4e63917462982d4b262468fea96ce0662