GkVulnerableImages: add "status" field to object identity
This makes the vulnerability status accessible as a filter in the DOOP metrics.

I am using a rather generic name here, rather than the more specific
"vulnerability_status" or "vuln_status". This label will now have to
appear on all DOOP metrics, so I want to use a name that might be useful
for other types of violations as well.