[Nova] Support secrets-injector
With the recent version bump we got a new helper in the `utils` chart
that makes it possible to support `secrets-injector` secrets will still
maintaining compatibility with not-yet-migrated secret values. In this
commit, we use this helper to make the `Nova` chart compatible with
`secrets-injector`.

We also bump the version of the `utils` chart again, because there is a
change to support resolving more usernames. It doesn't matter for us,
since we don't read them from Vault, yet, but we might in the future.

Since we implement custom helpers in `Nova` to support our cell2, we
also had to update them. For this, we use the available helpers from
other charts as far as possible so common functionality in generating
URLs is not duplicated.