[utils] Change kubernetes-entrypoint image repo
The old version of kubernetes-entrypoint was based on a rotten image and
not updated in years. We now switch to a new upstream and rebuild the
base image for security updates regularly through our
"shared-app-images" pipeline. The tag of the image contains the version
and a build number.

Since PCI forces us to do staged deployments for feature changes, we
track the version in the `utils` chart. This means we need a `utils`
chart bump to get a new version in, which will be transported through
the stages by the individual services' pipelines.

The build version, that's only incremented for base OS updates, can be
updated in all regions at the same time. Therefore, we track the latest
build version - and with that the one we want to deploy in a region -
through a region-local values file.

The idea would be to follow up with a pipeline that updates the build
version in the values file automatically every time a new image is
built.