Use StatefulSet instead of deployments
The deployment controller in case of `strategy: Recreate` terminates the
old pod and creates the new pod already. Because of our static IP
assignment via multus this would lead to duplicate addresses being used
while the old pod is not fully terminated and the new pod's networking
stack is already up. This did not pose a problem with IPv4 (aside from
it being still undesirable) but IPv6's duplicate address detection will
fail on the new pod, leading to it binding the address only tentatively.
This is of course undesired behavior. Hence we make the switch to
StatefulSets. Statefulsets make sure that there is only ever one
instance of a StatefulSets ordinal. By definition they are anyhow more
suiting to bird, as they maintain a stable identity and are tailored to
stateful applications.