Use different naming for statefulset
IPv6 is coming. So we need a new naming. Including the region in the
statefulset name has bothered me for quite some time as it provided
little value but instead bloats the pod name and requires the region to
be passed to everything.

The new naming now tells the operator what we are, a routeserver, which
address family we serve, which service number we do, and in which domain
we run.

We end up with

`routeserver-{{ $afi }}-service-{{ $service }}-domain-{{ $domain }}-{{ $instance }}`